packages/sardis-adk/pyproject.toml
[build-system]
requires = ["hatchling>=1.21"]
build-backend = "hatchling.build"

[project]
name = "sardis-adk"
version = "0.1.0"
description = "Google Agent Development Kit (ADK) integration for Sardis"
authors = [{name = "Sardis Team", email = "dev@sardis.sh"}]
maintainers = [{name = "Sardis Team", email = "dev@sardis.sh"}]
license = {text = "MIT"}
readme = "README.md"
requires-python = ">=3.11"
keywords = ["sardis", "adk", "google", "agent-development-kit", "payments", "ai-agents", "fintech"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Office/Business :: Financial",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Typing :: Typed",
]
dependencies = [
    "sardis>=0.3",
    "google-adk>=0.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
]

[project.urls]
Homepage = "https://sardis.sh"
Documentation = "https://sardis.sh/docs"
Repository = "https://github.com/EfeDurmaz16/sardis"
Changelog = "https://github.com/EfeDurmaz16/sardis/blob/main/CHANGELOG.md"
"Bug Tracker" = "https://github.com/EfeDurmaz16/sardis/issues"

[tool.hatch.metadata]
allow-direct-references = true

[tool.hatch.build.targets.wheel]
packages = ["src/sardis_adk"]
